14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,8 +7,8 @@
 int main(void)
 {
 	int i;
-	int num1 = 0, num2 = 1;
-	int num3  = num1 + num2;
+	long int num1 = 0, num2 = 1;
+	long int num3  = num1 + num2;
 
 	for (i = 1; i <= 50; i++)
 	{
